refactor: Move StaffFactory helper to common/
