+++ cms/djangoapps/contentstore/api/tests/base.py
@@ -6,8 +6,8 @@ Base test case for the course API views.
 from django.urls import reverse
 from rest_framework.test import APITestCase
 
+from common.djangoapps.student.tests.factories import StaffFactory
 from common.djangoapps.student.tests.factories import UserFactory
-from lms.djangoapps.courseware.tests.factories import StaffFactory
 from xmodule.modulestore.tests.django_utils import TEST_DATA_SPLIT_MODULESTORE, SharedModuleStoreTestCase
 from xmodule.modulestore.tests.factories import CourseFactory, ItemFactory
