+++ common/djangoapps/student/tests/test_roles.py
@@ -8,7 +8,7 @@ import six
 from django.test import TestCase
 from opaque_keys.edx.keys import CourseKey
 
-from lms.djangoapps.courseware.tests.factories import InstructorFactory, StaffFactory, UserFactory
+from lms.djangoapps.courseware.tests.factories import InstructorFactory, UserFactory
 from common.djangoapps.student.roles import (
     CourseBetaTesterRole,
     CourseInstructorRole,
@@ -20,6 +20,7 @@ from common.djangoapps.student.roles import (
     RoleCache
 )
 from common.djangoapps.student.tests.factories import AnonymousUserFactory
+from common.djangoapps.student.tests.factories import StaffFactory
 
 
 class RolesTestCase(TestCase):
