+++ lms/djangoapps/courseware/tests/test_entrance_exam.py
@@ -18,12 +18,13 @@ from lms.djangoapps.courseware.entrance_exams import (
 )
 from lms.djangoapps.courseware.model_data import FieldDataCache
 from lms.djangoapps.courseware.module_render import get_module, handle_xblock_callback, toc_for_course
-from lms.djangoapps.courseware.tests.factories import InstructorFactory, RequestFactoryNoCsrf, StaffFactory, UserFactory
+from lms.djangoapps.courseware.tests.factories import InstructorFactory, RequestFactoryNoCsrf, UserFactory
 from lms.djangoapps.courseware.tests.helpers import LoginEnrollmentTestCase
 from openedx.core.djangolib.testing.utils import get_mock_request
 from openedx.features.course_experience import DISABLE_COURSE_OUTLINE_PAGE_FLAG, DISABLE_UNIFIED_COURSE_TAB_FLAG
 from common.djangoapps.student.models import CourseEnrollment
 from common.djangoapps.student.tests.factories import AnonymousUserFactory, CourseEnrollmentFactory
+from common.djangoapps.student.tests.factories import StaffFactory
 from common.djangoapps.util.milestones_helpers import (
     add_course_content_milestone,
     add_course_milestone,
