+++ lms/djangoapps/courseware/tests/test_group_access.py
@@ -7,8 +7,9 @@ access control rules.
 import ddt
 from stevedore.extension import Extension, ExtensionManager
 
+from common.djangoapps.student.tests.factories import StaffFactory
 import lms.djangoapps.courseware.access as access
-from lms.djangoapps.courseware.tests.factories import StaffFactory, UserFactory
+from lms.djangoapps.courseware.tests.factories import UserFactory
 from xmodule.modulestore.django import modulestore
 from xmodule.modulestore.tests.django_utils import ModuleStoreTestCase
 from xmodule.modulestore.tests.factories import CourseFactory, ItemFactory
