+++ lms/djangoapps/courseware/tests/test_view_authentication.py
@@ -10,13 +10,13 @@ import pytz
 from django.urls import reverse
 
 from common.djangoapps.student.tests.factories import GlobalStaffFactory
+from common.djangoapps.student.tests.factories import StaffFactory
 from lms.djangoapps.courseware.access import has_access
 from lms.djangoapps.courseware.tests.factories import (
     BetaTesterFactory,
     InstructorFactory,
     OrgInstructorFactory,
     OrgStaffFactory,
-    StaffFactory
 )
 from lms.djangoapps.courseware.tests.helpers import CourseAccessTestMixin, LoginEnrollmentTestCase
 from openedx.features.enterprise_support.tests.mixins.enterprise import EnterpriseTestConsentRequired
