+++ openedx/core/djangoapps/discussions/tests/test_views.py
@@ -11,8 +11,8 @@ from rest_framework import status
 from rest_framework.test import APITestCase
 
 from common.djangoapps.student.tests.factories import GlobalStaffFactory
+from common.djangoapps.student.tests.factories import StaffFactory
 from common.djangoapps.student.tests.factories import UserFactory
-from lms.djangoapps.courseware.tests.factories import StaffFactory
 
 
 @unittest.skipUnless(settings.ROOT_URLCONF == 'lms.urls', 'URLs are only configured in LMS')
