+++ openedx/features/course_experience/tests/views/test_course_home.py
@@ -22,6 +22,7 @@ from waffle.testutils import override_flag
 from common.djangoapps.course_modes.models import CourseMode
 from common.djangoapps.course_modes.tests.factories import CourseModeFactory
 from common.djangoapps.student.tests.factories import GlobalStaffFactory
+from common.djangoapps.student.tests.factories import StaffFactory
 from common.djangoapps.util.date_utils import strftime_localized_html
 from lms.djangoapps.experiments.models import ExperimentData
 from lms.djangoapps.commerce.models import CommerceConfiguration
@@ -32,7 +33,6 @@ from lms.djangoapps.courseware.tests.factories import (
     InstructorFactory,
     OrgInstructorFactory,
     OrgStaffFactory,
-    StaffFactory
 )
 from lms.djangoapps.courseware.tests.helpers import get_expiration_banner_text
 from lms.djangoapps.courseware.utils import verified_upgrade_deadline_link
